refactor(zaino-state): co-locate FinalisedSource accumulator dispatch in the module
Move the two `FinalisedSource` accumulator forwarders
(`tx_out_set_info_accumulator_db`, `rebuild_tx_out_set_accumulator`) out of
finalised_source.rs into the gated `tx_out_set_accumulator` module, beside the
V1 impl they forward to. Inherent impls may live in any module of the defining
crate, and the private `require_v1` is reachable from the descendant module, so
this needs no visibility changes and folds the forwarders' `#[cfg]` into the
module gate (one fewer literal gettxoutsetinfo cfg). Compiler-verified in both
feature states.
